# perf(crypto): replace bignumber.js with native BigInt