--- a/__tests__/integration/core-api/handlers/wallets.test.ts
+++ b/__tests__/integration/core-api/handlers/wallets.test.ts
@@ -197,7 +197,7 @@ describe("API 2.0 - Wallets", () => {
                 const lockA = response.data.data[i];
                 const lockB = response.data.data[i + 1];
 
-                expect(Utils.BigNumber.make(lockA.amount).isGreaterThanOrEqualTo(lockB.amount)).toBeTrue();
+                expect(Utils.BigNumber.make(lockA.amount).isGreaterThanEqual(lockB.amount)).toBeTrue();
             }
         });
     });