--- a/packages/core-magistrate-crypto/src/interfaces.ts
+++ b/packages/core-magistrate-crypto/src/interfaces.ts
@@ -27,5 +27,5 @@ export interface IBridgechainUpdateAsset {
 }
 
 export interface IBridgechainResignationAsset {
-    bridgechainId: string;
+    bridgechainId: Utils.BigNumber;
 }