--- a/packages/core-transactions/src/handlers/transaction.ts
+++ b/packages/core-transactions/src/handlers/transaction.ts
@@ -56,7 +56,7 @@ export abstract class TransactionHandler implements ITransactionHandler {
             satoshiPerByte = 1;
         }
 
-        const transactionSizeInBytes: number = transaction.serialized.length / 2;
+        const transactionSizeInBytes: number = Math.round(transaction.serialized.length / 2);
         return Utils.BigNumber.make(addonBytes + transactionSizeInBytes).times(satoshiPerByte);
     }