--- a/packages/core-utils/src/delegate-calculator.ts
+++ b/packages/core-utils/src/delegate-calculator.ts
@@ -3,24 +3,31 @@ import { Blockchain, State } from "@arkecosystem/core-interfaces";
 import { Utils } from "@arkecosystem/crypto";
 import { supplyCalculator } from "./index";
 
+const toDecimal = (voteBalance: Utils.BigNumber, totalSupply: Utils.BigNumber): number => {
+    const decimals: number = 2;
+    const exponent: number = totalSupply.toString().length - voteBalance.toString().length + 4;
+
+    // @ts-ignore
+    const div = voteBalance.times(Math.pow(10, exponent)).dividedBy(totalSupply) / Math.pow(10, exponent - decimals);
+
+    return +Number(div).toFixed(2);
+};
+
 export const calculateApproval = (delegate: State.IWallet, height?: number): number => {
     if (!height) {
         height = app.resolvePlugin<Blockchain.IBlockchain>("blockchain").getLastBlock().data.height;
     }
 
-    const totalSupply = supplyCalculator.calculate(height);
+    const totalSupply = Utils.BigNumber.make(supplyCalculator.calculate(height));
     const voteBalance = Utils.BigNumber.make(delegate.getAttribute<Utils.BigNumber>("delegate.voteBalance"));
 
-    return +voteBalance
-        .times(100)
-        .dividedBy(totalSupply)
-        .toFixed();
+    return toDecimal(voteBalance, totalSupply);
 };
 
 export const calculateForgedTotal = (wallet: State.IWallet): string => {
     const delegate: State.IWalletDelegateAttributes = wallet.getAttribute("delegate");
-    const forgedFees: Utils.BigNumber = Utils.BigNumber.make(delegate.forgedFees);
-    const forgedRewards: Utils.BigNumber = Utils.BigNumber.make(delegate.forgedRewards);
+    const forgedFees: Utils.BigNumber = Utils.BigNumber.make(delegate.forgedFees || 0);
+    const forgedRewards: Utils.BigNumber = Utils.BigNumber.make(delegate.forgedRewards || 0);
 
     return forgedFees.plus(forgedRewards).toFixed();
 };