--- a/packages/core-webhooks/src/conditions.ts
+++ b/packages/core-webhooks/src/conditions.ts
@@ -8,14 +8,22 @@ const toBoolean = (value): boolean =>
         ? true
         : false;
 
+const compareBigNumber = (value, expected, comparison): boolean => {
+    try {
+        return Utils.BigNumber.make(value)[comparison](expected);
+    } catch {
+        return false;
+    }
+};
+
 export const between = (actual, expected): boolean => gt(actual, expected.min) && lt(actual, expected.max);
 export const contains = (actual, expected): boolean => actual.includes(expected);
 export const eq = (actual, expected): boolean => JSON.stringify(actual) === JSON.stringify(expected);
 export const falsy = (actual): boolean => actual === false || !toBoolean(actual);
-export const gt = (actual, expected): boolean => Utils.BigNumber.make(actual).isGreaterThan(expected);
-export const gte = (actual, expected): boolean => Utils.BigNumber.make(actual).isGreaterThanEqual(expected);
-export const lt = (actual, expected): boolean => Utils.BigNumber.make(actual).isLessThan(expected);
-export const lte = (actual, expected): boolean => Utils.BigNumber.make(actual).isLessThanEqual(expected);
+export const gt = (actual, expected): boolean => compareBigNumber(actual, expected, "isGreaterThan");
+export const gte = (actual, expected): boolean => compareBigNumber(actual, expected, "isGreaterThanEqual");
+export const lt = (actual, expected): boolean => compareBigNumber(actual, expected, "isLessThan");
+export const lte = (actual, expected): boolean => compareBigNumber(actual, expected, "isLessThanEqual");
 export const ne = (actual, expected): boolean => !eq(actual, expected);
 export const notBetween = (actual, expected): boolean => !between(actual, expected);
 export const regexp = (actual, expected): boolean => new RegExp(expected).test(actual);