--- a/packages/crypto/src/blocks/block.ts
+++ b/packages/crypto/src/blocks/block.ts
@@ -19,7 +19,10 @@ export class Block implements IBlock {
         }
 
         if (error) {
-            if (isException(value) || data.transactions.some((transaction: ITransactionData) => isException(transaction))) {
+            if (
+                isException(value) ||
+                data.transactions.some((transaction: ITransactionData) => isException(transaction))
+            ) {
                 // Validate again without bailing out on the first error to ensure that all properties get properly converted if necessary
                 validator.validateException("block", data);
             } else {
@@ -71,7 +74,7 @@ export class Block implements IBlock {
         const constants = configManager.getMilestone(data.height);
         const idHex: string = Block.getIdHex(data);
 
-        return constants.block.idFullSha256 ? idHex : BigNumber.make(idHex, 16).toFixed();
+        return constants.block.idFullSha256 ? idHex : BigNumber.make(idHex, 16).toString();
     }
 
     public serialized: string;
@@ -128,9 +131,9 @@ export class Block implements IBlock {
 
     public toJson(): IBlockJson {
         const data: IBlockJson = JSON.parse(JSON.stringify(this.data));
-        data.reward = this.data.reward.toFixed();
-        data.totalAmount = this.data.totalAmount.toFixed();
-        data.totalFee = this.data.totalFee.toFixed();
+        data.reward = this.data.reward.toString();
+        data.totalAmount = this.data.totalAmount.toString();
+        data.totalFee = this.data.totalFee.toString();
         data.transactions = this.transactions.map(transaction => transaction.toJson());
 
         return data;