--- a/packages/crypto/src/blocks/deserializer.ts
+++ b/packages/crypto/src/blocks/deserializer.ts
@@ -58,7 +58,7 @@ class Deserializer {
             block.previousBlock = block.previousBlockHex;
         } else {
             block.previousBlockHex = buf.readBytes(8).toString("hex");
-            block.previousBlock = BigNumber.make(block.previousBlockHex, 16).toString();
+            block.previousBlock = BigNumber.make(`0x${block.previousBlockHex}`).toString();
         }
 
         block.numberOfTransactions = buf.readUint32();