--- a/packages/crypto/src/blocks/serializer.ts
+++ b/packages/crypto/src/blocks/serializer.ts
@@ -1,4 +1,5 @@
 import ByteBuffer from "bytebuffer";
+import Long from "long";
 import { PreviousBlockIdFormatError } from "../errors";
 import { IBlockData, ITransactionData } from "../interfaces";
 import { configManager } from "../managers/config";
@@ -55,9 +56,9 @@ export class Serializer {
         buffer.writeUint32(block.height);
         buffer.append(block.previousBlockHex, "hex");
         buffer.writeUint32(block.numberOfTransactions);
-        buffer.writeUint64(+block.totalAmount.toFixed());
-        buffer.writeUint64(+block.totalFee.toFixed());
-        buffer.writeUint64(+block.reward.toFixed());
+        buffer.writeUint64(Long.fromString(block.totalAmount.toString()));
+        buffer.writeUint64(Long.fromString(block.totalFee.toString()));
+        buffer.writeUint64(Long.fromString(block.reward.toString()));
         buffer.writeUint32(block.payloadLength);
         buffer.append(block.payloadHash, "hex");
         buffer.append(block.generatorPublicKey, "hex");