--- a/packages/crypto/src/transactions/serializer.ts
+++ b/packages/crypto/src/transactions/serializer.ts
@@ -1,11 +1,12 @@
 /* tslint:disable:no-shadowed-variable */
 import ByteBuffer from "bytebuffer";
+import Long from "long";
 import { Utils } from "..";
 import { TransactionType, TransactionTypeGroup } from "../enums";
 import { TransactionVersionError } from "../errors";
 import { Address } from "../identities";
-import { ITransaction, ITransactionData } from "../interfaces";
 import { ISerializeOptions } from "../interfaces";
+import { ITransaction, ITransactionData } from "../interfaces";
 import { configManager } from "../managers/config";
 import { Base58, isSupportedTansactionVersion } from "../utils";
 import { TransactionTypeFactory } from "./types";
@@ -157,8 +158,8 @@ export class Serializer {
             }
         }
 
-        bb.writeInt64(+transaction.amount.toFixed());
-        bb.writeInt64(+transaction.fee.toFixed());
+        bb.writeInt64(Long.fromString(transaction.amount.toString()));
+        bb.writeInt64(Long.fromString(transaction.fee.toString()));
 
         if (assetSize > 0) {
             for (let i = 0; i < assetSize; i++) {
@@ -207,11 +208,11 @@ export class Serializer {
         } else {
             buffer.writeUint32(transaction.typeGroup);
             buffer.writeUint16(transaction.type);
-            buffer.writeUint64(+transaction.nonce);
+            buffer.writeUint64(Long.fromString(transaction.nonce.toString()));
         }
 
         buffer.append(transaction.senderPublicKey, "hex");
-        buffer.writeUint64(+transaction.fee);
+        buffer.writeUint64(Long.fromString(transaction.fee.toString()));
     }
 
     private static serializeVendorField(transaction: ITransaction, buffer: ByteBuffer): void {