--- a/packages/crypto/src/transactions/types/multi-payment.ts
+++ b/packages/crypto/src/transactions/types/multi-payment.ts
@@ -1,4 +1,5 @@
 import ByteBuffer from "bytebuffer";
+import Long from "long";
 import { TransactionType, TransactionTypeGroup } from "../../enums";
 import { IMultiPaymentItem, ISerializeOptions } from "../../interfaces";
 import { configManager } from "../../managers";
@@ -33,7 +34,7 @@ export class MultiPaymentTransaction extends Transaction {
         buffer.writeUint16(data.asset.payments.length);
 
         for (const p of data.asset.payments) {
-            buffer.writeUint64(+BigNumber.make(p.amount).toFixed());
+            buffer.writeUint64(Long.fromString(p.amount.toString()));
             buffer.append(Base58.decodeCheck(p.recipientId));
         }