--- a/packages/crypto/src/transactions/types/transfer.ts
+++ b/packages/crypto/src/transactions/types/transfer.ts
@@ -1,4 +1,5 @@
 import ByteBuffer from "bytebuffer";
+import Long from "long";
 import { TransactionType, TransactionTypeGroup } from "../../enums";
 import { ISerializeOptions } from "../../interfaces";
 import { Base58 } from "../../utils/base58";
@@ -24,7 +25,7 @@ export class TransferTransaction extends Transaction {
     public serialize(options?: ISerializeOptions): ByteBuffer {
         const { data } = this;
         const buffer: ByteBuffer = new ByteBuffer(24, true);
-        buffer.writeUint64(+data.amount);
+        buffer.writeUint64(Long.fromString(data.amount.toString()));
         buffer.writeUint32(data.expiration || 0);
         buffer.append(Base58.decodeCheck(data.recipientId));