--- a/packages/crypto/src/validation/keywords.ts
+++ b/packages/crypto/src/validation/keywords.ts
@@ -77,7 +77,16 @@ const bignumber = (ajv: Ajv) => {
                 const minimum = typeof schema.minimum !== "undefined" ? schema.minimum : 0;
                 const maximum = typeof schema.maximum !== "undefined" ? schema.maximum : "9223372036854775807"; // 8 byte maximum
 
-                const bignum = BigNumber.make(data);
+                if (data !== 0 && !data) {
+                    return false;
+                }
+
+                let bignum: BigNumber;
+                try {
+                    bignum = BigNumber.make(data);
+                } catch {
+                    return false;
+                }
 
                 if (parentObject && property) {
                     parentObject[property] = bignum;