<!--
Thanks for your interest in the project. Bugs filed and PRs submitted are appreciated!

Please make sure you're familiar with and follow the instructions in the [contributing guidelines](https://docs.ark.io/guidebook/contribution-guidelines/contributing.html).

Please fill out the information below to expedite the review and (hopefully) merge of your pull request!
-->

## Summary

<!-- What changes are being made? -->
Replaces `bignumber.js` with our own thin wrapper around the native [BigInt](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt) type. Depends on some yet unreleased `@arkecosystem/utils` fixes. With native BigInts the memory usage of my node is about 10% less after bootstrap compared to without this PR. 

Benchmarks before:
```
  new Block()
    ✓  fromData (0) x 7,887 ops/sec ±11.72% (84 runs sampled)
    ✓  fromData (150) x 57.06 ops/sec ±3.20% (62 runs sampled)
    ✓  fromData (150 Multipayments, 1MB) x 10.76 ops/sec ±1.59% (30 runs sampled)
    ✓  fromData (150 Multipayments, 150 Recipients each) x 2.62 ops/sec ±4.45% (11 runs sampled)
    ✓  fromData (150 Multipayments, 500 Recipients each) x 0.83 ops/sec ±0.74% (7 runs sampled)
    ✓  fromHex (0) x 9,721 ops/sec ±1.63% (90 runs sampled)
    ✓  fromHex (150) x 64.47 ops/sec ±1.14% (67 runs sampled)
  Block.serialize (0 transactions)
    ✓  core x 172,060 ops/sec ±1.03% (89 runs sampled)
  Block.serialize (150 transactions)
    ✓  core x 711 ops/sec ±1.42% (91 runs sampled)
  Block.deserialize (0 transactions)
    ✓  core x 31,361 ops/sec ±0.99% (89 runs sampled)
  Block.deserialize (150 transactions)
    ✓  core x 69.22 ops/sec ±1.12% (71 runs sampled)
  new Transaction (Type 0)
    ✓  fromData x 8,947 ops/sec ±0.95% (93 runs sampled)
    ✓  fromHex x 10,821 ops/sec ±0.78% (89 runs sampled)
    ✓  fromBytes x 10,700 ops/sec ±0.81% (94 runs sampled)
    ✓  fromBytesUnsafe x 132,489 ops/sec ±0.57% (90 runs sampled)
  Transaction.serialize (Type 0)
    ✓  core x 124,110 ops/sec ±1.19% (92 runs sampled)
  Transaction.deserialize (Type 0)
    ✓  core x 120,541 ops/sec ±0.87% (92 runs sampled)
```

After:
```
  new Block()
    ✓  fromData (0) x 11,483 ops/sec ±10.84% (91 runs sampled)
    ✓  fromData (150) x 61.69 ops/sec ±1.75% (65 runs sampled)
    ✓  fromData (150 Multipayments, 1MB) x 11.59 ops/sec ±0.88% (32 runs sampled)
    ✓  fromData (150 Multipayments, 150 Recipients each) x 2.84 ops/sec ±4.53% (12 runs sampled)
    ✓  fromData (150 Multipayments, 500 Recipients each) x 0.90 ops/sec ±0.40% (7 runs sampled)
    ✓  fromHex (0) x 12,402 ops/sec ±1.07% (89 runs sampled)
    ✓  fromHex (150) x 65.93 ops/sec ±1.97% (68 runs sampled)
  Block.serialize (0 transactions)
    ✓  core x 321,789 ops/sec ±1.46% (94 runs sampled)
  Block.serialize (150 transactions)
    ✓  core x 723 ops/sec ±1.31% (89 runs sampled)
  Block.deserialize (0 transactions)
    ✓  core x 59,453 ops/sec ±0.79% (92 runs sampled)
  Block.deserialize (150 transactions)
    ✓  core x 71.12 ops/sec ±1.14% (74 runs sampled)
  new Transaction (Type 0)
    ✓  fromData x 9,646 ops/sec ±0.72% (90 runs sampled)
    ✓  fromHex x 11,406 ops/sec ±0.59% (89 runs sampled)
    ✓  fromBytes x 11,316 ops/sec ±0.76% (92 runs sampled)
    ✓  fromBytesUnsafe x 159,295 ops/sec ±1.29% (87 runs sampled)
  Transaction.serialize (Type 0)
    ✓  core x 124,942 ops/sec ±1.24% (90 runs sampled)
  Transaction.deserialize (Type 0)
    ✓  core x 141,753 ops/sec ±1.07% (91 runs sampled)
```

<!-- Why are these changes necessary? -->

<!-- How were these changes implemented? -->

## Checklist

<!-- Have you done all of these things?  -->

- [ ] Documentation _(if necessary)_
- [ ] Tests _(if necessary)_
- [ ] Ready to be merged

<!-- Feel free to add additional comments. -->